implemented liquidity checker and advanced withdrawal steps
