+++ programs/dsla/src/state/reward.rs
@@ -1,9 +1,9 @@
 use anchor_lang::prelude::*;
 
-use super::DslaDecimal;
-
 #[account]
 pub struct Reward {
     last_claimed_period: u32,
-    period_reward: u32,
+    future_periods_reward: u128,
+    previous_periods_reward: u128,
+    current_period_reward: u128,
 }
